move new tests to new files
